Add files via upload
The dataset in Machine Learning-The German Credit Risk dataset. The goal is to predict if this loan credit would be a risk to the bank or not?

In simple terms, if the loan amount is given to the applicant, will he pay back or become a defaulter?

Loans form an integral part of banking operations. However, not all the loans are promptly paid and hence it is important for a bank to closely monitor its loan applications.

Since there are many applications which needs to be processed everyday, it will be helpful if there is a predictive model in place which can assist the executives to do their job by giving them a heads up about approval or rejection of a new loan application.